pyproject.toml
[tool.poetry]
name = "supportpulse-backend"
version = "0.1.0"
description = "SupportPulse AI - Real-time support intelligence platform powered by Confluent Kafka and Google Gemini"
authors = ["SupportPulse Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
confluent-kafka = "^2.6.0"
google-generativeai = "^0.8.0"
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
python-dotenv = "^1.0.1"
websockets = "^14.0"
structlog = "^24.4.0"
python-json-logger = "^3.2.0"
aiohttp = "^3.11.0"
asyncio = "^3.4.3"
uvloop = "^0.21.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
black = "^24.10.0"
ruff = "^0.8.0"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
